move WaitForServiceAccountInNamespace to e2e library
